+++ src/main/java/com/PedroA10/GeoAPI/service/UserService.java
@@ -2,6 +2,8 @@ package com.PedroA10.GeoAPI.service;
 
 import com.PedroA10.GeoAPI.dto.userdto.UserRequestDTO;
 import com.PedroA10.GeoAPI.dto.userdto.UserResponseDTO;
+import com.PedroA10.GeoAPI.exception.EmailAlreadyExistsException;
+import com.PedroA10.GeoAPI.exception.UserNotFoundException;
 import com.PedroA10.GeoAPI.mapper.UserMapper;
 import com.PedroA10.GeoAPI.model.User;
 import com.PedroA10.GeoAPI.repository.UserRepository;
@@ -10,7 +12,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UserService {
@@ -25,6 +26,10 @@ public class UserService {
   private PasswordEncoder passwordEncoder;
 
   public UserResponseDTO registerUser(UserRequestDTO requestDTO) {
+    if (userRepository.existsByEmail(requestDTO.getEmail())) {
+      throw new EmailAlreadyExistsException("Email already exists: " + requestDTO.getEmail());
+    }
+
     User users = userMapper.toEntity(requestDTO);
     users.setPassword(passwordEncoder.encode(users.getPassword()));
     User saveUser = userRepository.save(users);
@@ -32,16 +37,19 @@ public class UserService {
   }
 
   public List<UserResponseDTO> findAll() {
-    List<User> usersList =  userRepository.findAll();
-    return userMapper.toResponseDTOList(usersList);
+    return userMapper.toResponseDTOList(userRepository.findAll());
   }
 
-  public Optional<UserResponseDTO> findById(String id) {
-    Optional<User> usersOPT = userRepository.findById(id);
-    return usersOPT.map(userMapper::toResponseDTO);
+  public UserResponseDTO findById(String id) {
+    User user = userRepository.findById(id)
+      .orElseThrow(() -> new UserNotFoundException("User not found with id " + id));
+    return userMapper.toResponseDTO(user);
   }
 
   public void deleteUser(String id) {
+    if (!userRepository.existsById(id)) {
+      throw new UserNotFoundException("User not found with id " + id);
+    }
     userRepository.deleteById(id);
   }
 }
